figure 5, relative proportion of states (#251)
* updated figure7.svg

* renamed the figure5.py and figure7.py

* fixed a typo in the figure7.py comments

* switched plots on figure 6

* conflict

* temp

* changed figures

* Update figure6.py

* Update figure7.py
